refactor(runtime-contracts): keep child session keys beside the other key helpers
childSessionKey, childIdOf, and childAgentOf move into identifiers.ts and
read a key through parsedSessionKey instead of redeclaring the prefix,
separator, and segment. The two identical scalar schema property types in
acts become one.
