+++ packages/runtime-contracts/src/identifiers.ts
@@ -113,6 +113,29 @@ export function threadSessionKey(threadId: string, agent: AgentId = DEFAULT_AGEN
   );
 }
 
+/** A child's conversation: `agent:<agentId>:subagent:<childId>`, the id minted by the host. */
+export function childSessionKey(childId: string, agent: AgentId = DEFAULT_AGENT_ID): SessionKey {
+  if (!isIdentifier(childId) || childId.includes(SESSION_KEY_SEPARATOR)) {
+    throw new TypeError("child identifier must be a non-empty string without separators");
+  }
+  return sessionKey(
+    [SESSION_KEY_PREFIX, agent, SUBAGENT_SEGMENT, childId].join(SESSION_KEY_SEPARATOR),
+  );
+}
+
+/** The child id a key addresses, or nothing for a key of any other shape. */
+export function childIdOf(key: SessionKey | string): string | undefined {
+  const parsed = parsedSessionKey(key);
+  const [segment, childId] = parsed?.rest ?? [];
+  return parsed?.rest.length === 2 && segment === SUBAGENT_SEGMENT ? childId : undefined;
+}
+
+/** The agent a child key belongs to. */
+export function childAgentOf(key: SessionKey | string): AgentId | undefined {
+  const parsed = parsedSessionKey(key);
+  return parsed && parsed.rest[0] === SUBAGENT_SEGMENT ? agentId(parsed.agent) : undefined;
+}
+
 /**
  * The reversible encoding a provider's own identifiers take inside a session
  * key. A provider id or a provider's session id is untrusted text that may
