+++ packages/runtime-contracts/src/index.ts
@@ -12,11 +12,8 @@ export {
   type ChildSpawnReceipt,
   COMPLETION_DELIVERY_STATUS,
   type CompletionDeliveryStatus,
-  childAgentOf,
   childCompletionRecordFromWire,
-  childIdOf,
   childRunRecordFromWire,
-  childSessionKey,
   completionIdFor,
   isChildCleanup,
   isChildContextMode,
@@ -87,6 +84,9 @@ export {
   CONVERSATION_KIND,
   type ConversationKind,
   callId,
+  childAgentOf,
+  childIdOf,
+  childSessionKey,
   conversationKindOf,
   DEFAULT_AGENT_ID,
   type DeliveryId,
